%[ER: 182]%
Removed DH_Empty

+++ CEP/Tools/AccepTest1/src/DH_CorrCube.h
@@ -26,7 +26,7 @@ namespace LOFAR
 class DH_CorrCube: public DataHolder
 {
 public:
-  typedef complex<float> BufferType;
+  typedef complex<short> BufferType;
 
   explicit DH_CorrCube (const string& name, 
 			const int stations,
